requirements.txt
streamlit
tensorflow
keras
opencv-python
numpy
pillow